+++ src/game.cpp
@@ -35,7 +35,7 @@ void Game::createWindow(SDL_WindowFlags flag) {
         throw std::runtime_error("HGD-1000: Error creating window => SDL_Init failed: " + std::string(SDL_GetError()));
     }
 
-    window = SDL_CreateWindow("Geometry Dash 1.0 (1.0.5)", x, y, w, h, flag);
+    window = SDL_CreateWindow("Geometry Dash 1.0 (1.0.7)", x, y, w, h, flag);
     if (!window) {
         SDL_Quit();
         throw std::runtime_error("HGD-1001: Error creating window => SDL_CreateWindow failed: " + std::string(SDL_GetError()));
@@ -151,6 +151,7 @@ void Game::run() {
         if (timePerFrame >= frameTime) SDL_Delay(timePerFrame - frameTime);
         if (SDL_GetTicks() - fpsTimer >= 1000) {
             currentFPS = frameCounter / ((SDL_GetTicks() - fpsTimer) / 1000.0f);
+            // std::cout << "FPS: " << currentFPS << std::endl;
             frameCounter = 0;
             fpsTimer = SDL_GetTicks();
         }